requirements.txt
Django>=2.1.7
psycopg2
python-decouple
whitenoise
django-spa
djangorestframework
djangorestframework-simplejwt
django-cors-headers
Pillow